# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  v1.AddCommentRequest:
    properties:
      content:
        type: string
      mentioned_user_ids:
        items:
          type: string
        type: array
    required:
    - content
    type: object
...
  v1.CreateContactRequest:
    properties:
      company:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
    required:
    - name
    - phone
    type: object
...
  v1.UpdateContactRequest:
    properties:
      assigned_to:
        type: string
      company:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@crm.com
    name: CRM Support
  description: API untuk sistem CRM (Customer Relationship Management)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CRM Project API
  version: "1.0"
paths:
...
  /analytics/forecasting:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get analytics forecasting
      tags:
      - Dashboard
...
  /contacts:
    get:
      description: Ambil daftar contacts dengan filter dan pagination
      parameters:
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah per halaman (default 10)
        in: query
        name: limit
        type: integer
      - description: Kata kunci pencarian
        in: query
        name: search
        type: string
      - description: Filter source (whatsapp/manual)
        in: query
        name: source
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Buat contact baru ke sistem
      parameters:
      - description: Data contact
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat contact baru
      tags:
      - contacts
  /contacts/{id}:
    delete:
      description: Soft delete contact berdasarkan ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus contact
      tags:
      - contacts
    get:
      description: Ambil detail contact berdasarkan ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail contact
      tags:
      - contacts
    patch:
      consumes:
      - application/json
      description: Update data contact secara partial
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diupdate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update contact
      tags:
      - contacts
  /contacts/{id}/restore:
    patch:
      description: Restore contact yang sudah di-soft delete
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Restore contact
      tags:
      - contacts
  /contacts/export:
    get:
      description: Download daftar contacts dalam format CSV
      parameters:
      - description: Filter source (whatsapp/manual)
        in: query
        name: source
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Export contacts ke CSV
      tags:
      - contacts
  /contacts/import:
    post:
      consumes:
      - multipart/form-data
      description: Upload file CSV untuk import contacts secara massal
      parameters:
      - description: File CSV
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Import contacts dari CSV
      tags:
      - contacts
  /contacts/trash:
    get:
      description: Ambil daftar contact yang sudah di-soft delete
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Contacts yang dihapus
      tags:
      - contacts
  /dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get statistik dashboard
      tags:
      - Dashboard
...
  /deals/{id}/comments:
    get:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List komentar deal
      tags:
      - Deal Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: string
      - description: Data komentar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tambah komentar ke deal
      tags:
      - Deal Comments
  /deals/{id}/comments/{commentId}:
    delete:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus komentar deal
      tags:
      - Deal Comments
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"